Update CMakeLists structure and add project metadata
This commit restructures the package CMakeLists.txt and creates a new src/CMakeLists.txt file with more detailed categorization of source and header files. This restructure offers clearer maintenance and makes it easier for raygui_cpp contributors to find and alter specific files. Furthermore, this commit adds basic project metadata such as version number and description to the root CMakeLists.txt file.
The library is compiled as a static one.