Add ToDo list in README and correct GuiIsLocked return type
An 'Implementation todo list' section was added to README.md to keep track of objects that are yet to be implemented.
Also, in the Globals class, 'GuiIsLocked()' function was incorrectly void, and it was modified to return a boolean value so that it can be used logically in further implementation.