Add method to create Bounds of given size
A new static method, `OfSize`, has been created in Bounds.h. This method can be used to create a Bounds object of specified width and height with a default position of (0,0). This change is made to provide a more intuitive way of creating Bounds objects without the need to set position explicitly. This can be used when positioning components relatively to other ones and to avoid specifying the (0,0) position.