Add support for CPM and remove unnecessary raylib.h and implement raygui
Support for CPM is added again with better integration of the source code. Due to this change, raylib.h is no longer needed in the project because it is imported by CPM. Also, the implementation of raygui is added only once in the main header file of the project to prevent redefinition errors thrown by the compiler.