Fixed child-parent relation and introduce Component utility function
Simplified the button generation code in the 'ChildComponents.cpp' example by reusing the button bounds, and reduced the number of buttons for clarity. Several child configurations were added to visualize the new functionality. Introduced a utility function 'ToComponent' to improve code readability when a cast to parent class is needed.