Add onClick and onUpdate callbacks to Button component
Extended the Button and Component classes to allow passing callback functions for click and update events. The click event fires when the button is clicked, and the update event fires during each component update. This feature was demonstrated in a new example called "ButtonCallbacks.cpp", which was added to the list of examples.