Add state handling and enable/disable functionality to components
Implemented a new state handling mechanism using a macro, WITH_STATE_RENDER. This allows for the GUI state to be properly set before rendering a component. This is particularly beneficial for supporting enabling and disabling the interactive components. Enabled and disabled methods were added to the Component class to manage a new 'm_enabled' flag, which affects onclick behavior as well.